+++ components/translation/translation.js
@@ -1,0 +1,1 @@
+angular.module( 'gj.Translation', [ 'gj.Model' ] );